flask/requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
CacheControl==0.12.6
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==0.9.6
MarkupSafe==1.1.1
meraki==1.3.0
meraki-sdk==1.5.0
msgpack==1.0.0
multidict==5.0.2
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.25.0
six==1.15.0
ttp==0.5.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
gunicorn==20.0.4
pymongo==3.11.1
python-dotenv==0.15.0
